Closes: #13 Reorder Folder Structure and Finalize API
