+++ RealTime/.gitignore
@@ -1,2 +1,6 @@
 # Ignoring all virtual env files in env folder
 env/
+
+# Ignoring all pycache files
+__pycache__/
+*.pyc
