+++ RealTime/Controllers/server.py
@@ -3,6 +3,11 @@ import datetime
 from geopy.geocoders import Nominatim
 import urllib.request
 import json
+import sys
+
+sys.path.append('../')
+
+from Utils.utils import create_payload
 
 app = Flask(__name__)
 
@@ -11,32 +16,6 @@ geolocator = Nominatim(user_agent='WeatherDashboard')
 
 # Base url of API for fetching all real time weather details.
 url = "https://api.weather.gov/points/"
-
-'''
-    Function to convert bytes to json type.
-'''
-def bytesToJSON(data):
-    my_json = data.decode('utf8').replace("'", '"')
-    new_d = json.loads(my_json)
-    return json.dumps(new_d)
-
-'''
-    Function to fetch the forecast details from API.
-'''
-# @app.route('/forecast/<data>', methods=['GET'])
-def forecast(forecastURL):
-    data = urllib.request.urlopen(forecastURL).read()
-    json_data = bytesToJSON(data)
-    return json_data
-
-'''
-    Function to generate the payload for user consumption.
-'''
-def create_payload(data):
-    json_data = bytesToJSON(data)
-    payload = { }
-    payload['forecast'] = forecast(json_data.properties.forecast)
-    return jsonify(payload)
 
 '''
     Route to get the weather forecast from the city name given by the user.
